Apply payment by Ngan Luong

+++ SaleShop.Web/Assets/client/js/controllers/shoppingCart.js
@@ -94,6 +94,20 @@
                 }
             }
         });
+
+        $('input[name="paymentMethod"]').off('click').on('click', function () {
+            if ($(this).val() == 'NL') {
+                $('.boxContent').hide();
+                $('#nganluongContent').show();
+            }
+            else if ($(this).val() == 'ATM_ONLINE') {
+                $('.boxContent').hide();
+                $('#bankContent').show();
+            }
+            else {
+                $('.boxContent').hide();
+            }
+        });
     },
     getLoginUser:function() {
         $.ajax({
@@ -119,7 +133,8 @@
             CustomerMessage: $('#txtMessage').val(),
             CustomerMobile: $('#txtPhone').val(),
             CustomerName: $('#txtName').val(),
-            PaymentMethod:'Thanh toán tiền mặt',
+            PaymentMethod: $('input[name="paymentMethod"]:checked').val(),
+            BankCode: $('input[groupname="bankcode"]:checked').prop('id'),
             Status:false
         };
 
@@ -132,15 +147,22 @@
             dataType: 'json',
             success: function (res) {
                 if (res.status) {
-                    $('#divCheckout').hide();
-                    cart.deleteAll();
 
-                    setTimeout(function() {
-                            $('#cartContent').html('Cảm ơn bạn đã đặt hàng thành công. Chúng tôi sẽ liên hệ sớm nhất');
-                        },
-                        2000);
+                    if (res.urlCheckout != undefined && res.urlCheckout != '') {
+
+                        window.location.href = res.urlCheckout;
+                    } else {
+                        $('#divCheckout').hide();
+                        cart.deleteAll();
+
+                        setTimeout(function () {
+                                $('#cartContent').html('Cảm ơn bạn đã đặt hàng thành công. Chúng tôi sẽ liên hệ sớm nhất');
+                            },
+                            2000);
+                    }
                 } else {
-                    alert(res.message);
+                    $('#divMessage').show();
+                    $('#divMessage').text(res.message);
                 }
             }
         });
